finish HW3

+++ number.h
@@ -1,12 +1,12 @@
 #ifndef NUMBER_H
 #define NUMBER_H
 
-#include "Term.h"
+#include "term.h"
 using namespace std;
 
 class Number : public Term {
 public:
-	Number(int x) { temp << x; _symbol = temp.str();type = "Number"; }
+	Number(double x) { temp << x; _symbol = temp.str();type = "Number"; }
 	template <class Type>
 	bool match(Type &compare) {
 		if (compare.type == "Variable")
